# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Greenlight API documentation",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:4000",
    "basePath": "/v1",
    "paths": {
...
        "/tokens/authentication": {
            "post": {
                "description": "Generate a new token for a user using the given credentials",
                "tags": [
                    "Token"
                ],
                "summary": "Get user authentication token",
                "parameters": [
                    {
                        "description": "auth token credentials",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AuthTokenRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/commons.ResponseObject"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.TokenResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "e.g. status: error, message: the error reason",
                        "schema": {
                            "$ref": "#/definitions/commons.ResponseObject"
                        }
                    },
                    "422": {
                        "description": "status: fail",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/commons.ResponseObject"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.ValidationError"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "e.g. status: error, message: the error reason",
                        "schema": {
                            "$ref": "#/definitions/commons.ResponseObject"
                        }
                    }
                }
            }
        },
        "/users": {
            "post": {
                "description": "create a new user with given details",
                "tags": [
                    "Users"
                ],
                "summary": "Create new user",
                "parameters": [
                    {
                        "description": "create user",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/commons.ResponseObject"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.SingleUserResponse"
                                        }
                                    }
                                }
                            ]
                        },
                        "headers": {
                            "Location": {
                                "type": "string",
                                "description": "/v1/users/QbPy4B7a2Lw1Kg7ogoEWj9k3NGMRVY"
                            }
                        }
                    },
                    "400": {
                        "description": "e.g. status: error, message: the error reason",
                        "schema": {
                            "$ref": "#/definitions/commons.ResponseObject"
                        }
                    },
                    "422": {
                        "description": "status: fail",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/commons.ResponseObject"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.ValidationError"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "e.g. status: error, message: the error reason",
                        "schema": {
                            "$ref": "#/definitions/commons.ResponseObject"
                        }
                    }
                }
            }
        },
        "/users/activated": {
            "put": {
                "description": "activate the account of a user using the given token",
                "tags": [
                    "Users"
                ],
                "summary": "Activate user account",
                "parameters": [
                    {
                        "description": "activate user",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ActivateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/commons.ResponseObject"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.SingleUserResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "e.g. status: error, message: the error reason",
                        "schema": {
                            "$ref": "#/definitions/commons.ResponseObject"
                        }
                    },
                    "409": {
                        "description": "e.g. status: error, message: unable to update the record due to an edit conflict, please try again",
                        "schema": {
                            "$ref": "#/definitions/commons.ResponseObject"
                        }
                    },
                    "422": {
                        "description": "status: fail",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/commons.ResponseObject"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.ValidationError"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "e.g. status: error, message: the error reason",
                        "schema": {
                            "$ref": "#/definitions/commons.ResponseObject"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "commons.ResponseObject": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "description": "(success|fail|error)",
                    "type": "integer"
                }
            }
        },
...
        "dto.ActivateUserRequest": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
        "dto.AuthTokenRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "dto.CreateUserRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "description": "unique email address",
                    "type": "string"
                },
                "name": {
                    "description": "fullname",
                    "type": "string"
                },
                "password": {
                    "description": "minimum 8 bytes maximum 72 bytes",
                    "type": "string"
                }
            }
        },
...
        "dto.SingleUserResponse": {
            "type": "object",
            "properties": {
                "user": {
                    "$ref": "#/definitions/dto.UserResponse"
                }
            }
        },
        "dto.Token": {
            "type": "object",
            "properties": {
                "expiry": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
        "dto.TokenResponse": {
            "type": "object",
            "properties": {
                "authentication_token": {
                    "$ref": "#/definitions/dto.Token"
                }
            }
        },
        "dto.UserResponse": {
            "type": "object",
            "properties": {
                "activated": {
                    "type": "boolean"
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "version": {
                    "type": "integer"
                }
            }
        },
        "dto.ValidationError": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        }
    }
}